External Web Update
* cropped LifeToTheFullest Images
* fixes Button hover for Hero section
* updated Start here links nav
* fixed highlight effect for LifeToTheFullest section
* added Card scaling on hover for ItAllStartsHere section
* fixed background ellipse in LifeToTheFullest section
* refactored `external-home` into components under `components/ExternalHome`
