+++ components/ExternalHome/index.js
@@ -1,0 +1,10 @@
+import DoMoreThan from './DoMoreThan';
+import LifeIsCrazy from './LifeIsCrazy';
+import LifeToTheFullest from './LifeToTheFullest';
+import StartHere from './StartHere';
+
+/**
+ * These are components apart of the External Landing Home page.
+ */
+
+export { DoMoreThan, LifeIsCrazy, LifeToTheFullest, StartHere };
